+++ src/test/core/test_core.py
@@ -31,7 +31,8 @@ class TestCore(TestBase):
 
     LOREM_FILENAME = 'lorem.txt'
     SIMPLE_REPORT_JSON = 'simple_report_expected.json'
-    SIMPLE_REPORT_MD5 = '37caf581f36b1450900d6d770f9c497e'
+    SIMPLE_REPORT_UPDATE_JSON = 'simple_report_for_update.json'
+    SIMPLE_REPORT_MD5 = 'bba1ed174db424cc0dcd10766a28b2e3'
     SIMPLE_CONFIG_MD5 = 'ab4b71b790f2b12aa802b8eaa1658951'
 
     class mock_args:
@@ -487,7 +488,7 @@ class TestMainScript(TestCore):
         with open(ini_path, 'w') as out_file:
             print(ini_string, file=out_file)
         # run djerba.py and check the results
-        json_path = os.path.join(self.test_source_dir, self.SIMPLE_REPORT_JSON)
+        json_path = os.path.join(self.test_source_dir, self.SIMPLE_REPORT_UPDATE_JSON)
         cmd = [
             'djerba.py', mode,
             '--work-dir', work_dir,
@@ -501,7 +502,7 @@ class TestMainScript(TestCore):
         html_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.html')
         with open(html_path) as html_file:
             html_string = html_file.read()
-        self.assert_report_MD5(html_string, '0be6f09432ed2e861af63565d8205716')
+        self.assert_report_MD5(html_string, '97070aad9d0e9d9b05564a93ecab9fc3')
         pdf_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.pdf')
         self.assertTrue(os.path.isfile(pdf_path))
         # again, with the --write-json option
@@ -518,7 +519,7 @@ class TestMainScript(TestCore):
         work_dir = self.tmp_dir
         summary_path = os.path.join(self.test_source_dir, 'alternate_summary.txt')
         # run djerba.py and check the results
-        json_path = os.path.join(self.test_source_dir, self.SIMPLE_REPORT_JSON)
+        json_path = os.path.join(self.test_source_dir, self.SIMPLE_REPORT_UPDATE_JSON)
         cmd = [
             'djerba.py', mode,
             '--work-dir', work_dir,
@@ -532,7 +533,7 @@ class TestMainScript(TestCore):
         html_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.html')
         with open(html_path) as html_file:
             html_string = html_file.read()
-        self.assert_report_MD5(html_string, 'a0a3308b8ed6dd4dacb7bfff8f7adc72')
+        self.assert_report_MD5(html_string, '97070aad9d0e9d9b05564a93ecab9fc3')
         pdf_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.pdf')
         self.assertTrue(os.path.isfile(pdf_path))
         # again, with the --write-json option